# Conflicts:
#	External/Plugins/SourceControl/Sources/Git/AddCommand.cs
#	External/Plugins/SourceControl/Sources/Git/CommitCommand.cs
#	External/Plugins/SourceControl/Sources/Git/MenuItems.cs
#	External/Plugins/SourceControl/Sources/Git/MoveCommand.cs
#	External/Plugins/SourceControl/Sources/Git/RenameCommand.cs
#	External/Plugins/SourceControl/Sources/Git/Status.cs
#	External/Plugins/SourceControl/Sources/Mercurial/AddCommand.cs
#	External/Plugins/SourceControl/Sources/Mercurial/CommitCommand.cs
#	External/Plugins/SourceControl/Sources/Mercurial/MercurialManager.cs
#	External/Plugins/SourceControl/Sources/Mercurial/MoveCommand.cs
#	External/Plugins/SourceControl/Sources/Mercurial/RenameCommand.cs
#	External/Plugins/SourceControl/Sources/Mercurial/Status.cs
#	External/Plugins/SourceControl/Sources/Subversion/AddCommand.cs
#	External/Plugins/SourceControl/Sources/Subversion/CommitCommand.cs
#	External/Plugins/SourceControl/Sources/Subversion/MoveCommand.cs
#	External/Plugins/SourceControl/Sources/Subversion/RenameCommand.cs
#	External/Plugins/SourceControl/Sources/Subversion/Status.cs

+++ PluginCore/SevenZip/Compress/RangeCoder/RangeCoder.cs
@@ -1,4 +1,3 @@
-using System;
 using System.IO;
 
 namespace SevenZip.Compression.RangeCoder
@@ -9,7 +8,7 @@ namespace SevenZip.Compression.RangeCoder
 
         Stream Stream;
 
-        public UInt64 Low;
+        public ulong Low;
         public uint Range;
         uint _cacheSize;
         byte _cache;
@@ -65,7 +64,7 @@ namespace SevenZip.Compression.RangeCoder
 
         public void ShiftLow()
         {
-            if ((uint)Low < (uint)0xFF000000 || (uint)(Low >> 32) == 1)
+            if ((uint)Low < 0xFF000000 || (uint)(Low >> 32) == 1)
             {
                 byte temp = _cache;
                 do
